backend/requirements.txt
django
djangorestframework
mysqlclient
django-cors-headers
requests
djangorestframework-simplejwt